Moving namespace of SpApiConfig
